pyproject.toml
[tool.poetry]
name = "foreninglet-data"
version = "0.1.0"
description = "A way to retrieve information about members from the ForeningLet API (www.foreninglet.dk) - ForeningLet is a member system for small (typically) voluntary organizations in Denmark"
authors = ["Carsten Skov <carsten@simcax.dk>"]
readme = "README.md"
packages = [{include = "foreninglet_data"}]

[tool.poetry.dependencies]
python = "^3.11"
dataclasses-json = "^0.5.7"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^23.7.0", allow-prereleases = true}
dagger-io = "^0.9.5"
typer = "^0.9.0"
faker = "^19.2.0"
pytest-vcr = "^1.0.2"
python-dotenv = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
